Provide example initializer
